make types file and helper file for queryResults/Error
