+++ server/titles.ts
@@ -1,0 +1,5 @@
+export const SQLtitles = {
+    booleanBased: 'Boolean Based SQL Injection',
+    errorBased: 'Error Based SQL Injection',
+    timeBased: 'Time-Based Blind SQL Injection',
+  };
